Fix execution result visualization problem
Execution result wasn't been shown due to a binder queue name problem.
Main window in window.zul was associated to a named queue in order to
allow SourceCodeComposer to listen the global events. As a result, calls
to BindUtils.notifyChange() done in MainViewModel with "null" queue name
stopped to work.